Creating a Accordian React App Part 1
