+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "bootstrap": "^5.2.3",
     "create-react-app": "^5.0.1",
     "react": "^18.2.0",
     "react-digital-clock": "^0.1.2",
