+++ src/components/accordian/Accordian.jsx
@@ -1,0 +1,16 @@
+import React, { useState } from "react";
+import { questions } from "./api";
+import "./accordian.css";
+
+const Accordian = () => {
+	const [data, setData] = useState(questions);
+	return (
+		<>
+			{data.map((curr) => {
+				return curr;
+			})}
+		</>
+	);
+};
+
+export default Accordian;
