+++ src/components/accordian/api.jsx
@@ -1,0 +1,12 @@
+export const questions = [
+	{
+		id: 1,
+		question: "What is virtual DOM ?",
+		answer: "DOM stands for....",
+	},
+	{
+		id: 2,
+		question: "How do you create a React app ?",
+		answer: "Install the create-react-app....",
+	},
+];
